pyproject.toml
[project]
name = "langchain-chat-with-your-data"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "chromadb>=0.5.20",
    "kwwutils>=0.1.15",
    "langchain>=0.3.14",
    "pypdf>=5.1.0",
    "sentence-transformers>=3.3.1",
    "tf-keras>=2.18.0",
]

[project.optional-dependencies]
dev = [
    "pytest-cov>=5.0.0",
    "pytest-datadir>=1.5.0",
    "pytest-flake8>=1.1.1",
    "pytest-instafail>=0.5.0",
    "pytest-json>=0.4.0",
    "pytest-leaks>=0.3.1",
    "pytest-pycodestyle>=2.3.1",
    "pytest-redis>=3.0.2",
    "pytest-regressions>=2.5.0",
    "pytest-repeat>=0.9.3",
    "pytest-replay>=1.5.1",
    "pytest-rerunfailures>=14.0",
    "pytest-selenium>=4.1.0",
    "pytest-sugar>=1.0.0",
    "pytest-timeout>=2.3.1",
    "pytest-vagrant>=2.1.0",
    "pytest-vcr>=1.0.2",
    "pytest-xdist>=3.5.0",
    "pytest-xvfb>=3.0.0",
]
